Changed the range to `u8`
The benchmarks perform the same but the memory usage should be
lower.

+++ src/distr/other.rs
@@ -154,11 +154,11 @@ impl Distribution<u8> for Alphanumeric {
 
 impl Distribution<u8> for Alphabetic {
     fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> u8 {
-        const RANGE: u32 = 26 + 26;
+        const RANGE: u8 = 26 + 26;
         const GEN_ASCII_STR_CHARSET: &[u8] = b"ABCDEFGHIJKLMNOPQRSTUVWXYZ\
                 abcdefghijklmnopqrstuvwxyz";
 
-        GEN_ASCII_STR_CHARSET[rng.random_range(0..RANGE as usize)]
+        GEN_ASCII_STR_CHARSET[rng.random_range(0..RANGE) as usize]
     }
 }
